refactor: extract shared types and utils
