+++ .gitignore
@@ -69,6 +69,7 @@ jspm_packages/
 # Nuxt.js build / generate output
 .nuxt
 dist
+tests/dist
 
 # Gatsby files
 .cache/
