+++ client/src/app/components/WorkflowCard.tsx
@@ -1,23 +1,7 @@
 "use client";
 
-interface WorkflowRun {
-  id: number;
-  name: string;
-  status: string;
-  conclusion: string | null;
-  html_url: string;
-  created_at: string;
-  updated_at: string;
-}
-
-interface WorkflowWithLatestRun {
-  workflow_id: number;
-  workflow_name: string;
-  workflow_path: string;
-  workflow_state: string;
-  workflow_url: string;
-  latest_run: WorkflowRun | null;
-}
+import type { WorkflowWithLatestRun } from '@/types/github';
+import { formatRelativeTime } from '@/utils/time';
 
 interface WorkflowCardProps {
   workflow: WorkflowWithLatestRun;
@@ -46,32 +30,6 @@ export default function WorkflowCard({ workflow }: WorkflowCardProps) {
     }
   };
 
-  const formatRelativeTime = (dateString: string) => {
-    const date = new Date(dateString);
-    const now = new Date();
-    const diffInSeconds = Math.floor((now.getTime() - date.getTime()) / 1000);
-    const diffInMinutes = Math.floor(diffInSeconds / 60);
-    const diffInHours = Math.floor(diffInMinutes / 60);
-    const diffInDays = Math.floor(diffInHours / 24);
-
-    const timeFormat = date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-
-    if (diffInSeconds < 60) {
-      return `Just now`;
-    } else if (diffInMinutes < 60) {
-      return `${diffInMinutes} minute${diffInMinutes !== 1 ? 's' : ''} ago`;
-    } else if (diffInHours < 24) {
-      return `${diffInHours} hour${diffInHours !== 1 ? 's' : ''} ago`;
-    } else if (diffInDays === 0) {
-      return `Today at ${timeFormat}`;
-    } else if (diffInDays === 1) {
-      return `Yesterday at ${timeFormat}`;
-    } else if (diffInDays < 7) {
-      return `${diffInDays} day${diffInDays !== 1 ? 's' : ''} ago`;
-    } else {
-      return `${date.toLocaleDateString()} at ${timeFormat}`;
-    }
-  };
 
   return (
     <div
