+++ client/src/app/dashboard/page.tsx
@@ -8,37 +8,7 @@ import FilterPanel from '../components/FilterPanel';
 import useAuth from '@/hooks/useAuth';
 import { apiRequest } from '@/utils/api';
 import Head from 'next/head';
-
-interface Repo {
-  id: number;
-  name: string;
-  full_name: string;
-  private: boolean;
-  owner?: {
-    login: string;
-    type: 'User' | 'Organization';
-  };
-  updated_at: string;
-}
-
-interface WorkflowRun {
-  id: number;
-  name: string;
-  status: string;
-  conclusion: string | null;
-  html_url: string;
-  created_at: string;
-  updated_at: string;
-}
-
-interface WorkflowWithLatestRun {
-  workflow_id: number;
-  workflow_name: string;
-  workflow_path: string;
-  workflow_state: string;
-  workflow_url: string;
-  latest_run: WorkflowRun | null;
-}
+import type { Repo, WorkflowRun, WorkflowWithLatestRun } from '@/types/github';
 
 function DashboardContent() {
   const { authStatus, loading: authLoading } = useAuth();
