+++ client/src/utils/time.ts
@@ -1,0 +1,26 @@
+export function formatRelativeTime(dateString: string): string {
+  const date = new Date(dateString);
+  const now = new Date();
+  const diffInSeconds = Math.floor((now.getTime() - date.getTime()) / 1000);
+  const diffInMinutes = Math.floor(diffInSeconds / 60);
+  const diffInHours = Math.floor(diffInMinutes / 60);
+  const diffInDays = Math.floor(diffInHours / 24);
+
+  const timeFormat = date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
+
+  if (diffInSeconds < 60) {
+    return 'Just now';
+  } else if (diffInMinutes < 60) {
+    return `${diffInMinutes} minute${diffInMinutes !== 1 ? 's' : ''} ago`;
+  } else if (diffInHours < 24) {
+    return `${diffInHours} hour${diffInHours !== 1 ? 's' : ''} ago`;
+  } else if (diffInDays === 0) {
+    return `Today at ${timeFormat}`;
+  } else if (diffInDays === 1) {
+    return `Yesterday at ${timeFormat}`;
+  } else if (diffInDays < 7) {
+    return `${diffInDays} day${diffInDays !== 1 ? 's' : ''} ago`;
+  } else {
+    return `${date.toLocaleDateString()} at ${timeFormat}`;
+  }
+}
