+++ tests/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "ES2019",
+    "esModuleInterop": true,
+    "moduleResolution": "node",
+    "outDir": "dist",
+    "lib": ["ES2019", "DOM"]
+  },
+  "include": ["../client/src/utils/time.ts"],
+  "exclude": ["dist"]
+}
